requirements.txt
pycryptodome
tinyec